+++ deovi/scrapper.py
@@ -5,9 +5,11 @@ from pathlib import Path
 
 from deepdiff import DeepDiff
 
-from tmdbv3api import Configuration, TMDb, TV
+from tmdbv3api import Configuration, TMDb, TV, Movie
 
 import yaml
+
+from .utils.jsons import ExtendedJsonEncoder
 
 
 class TmdbScrapper:
@@ -30,7 +32,11 @@ class TmdbScrapper:
         poster_size (string): Size name as supported from TMDb API.
         poster_filename (string): Filename to use to write download poster image,
             without any extension.
-        dry (boolean): If enabled nothing will be written or removed.
+        dry (boolean): If enabled nothing will be written or removed. The JSON payload
+            from the ``debug`` is always written no matter of the dry option.
+        debug (boolean): If enabled the fetched payload from TmdbScrapper (not to
+            confuse with the real TMDB payload) is saved on disk in a JSON file named
+            after the media tmdb_id, file is saved in the current working directory.
     """
     DEFAULT_LANGUAGE = "fr"
     DEFAULT_POSTER_SIZE = "w780"
@@ -38,8 +44,9 @@ class TmdbScrapper:
     DEFAULT_MANIFEST_FORMAT = "yaml"
 
     def __init__(self, api_key, language=None, poster_size=None,
-                 poster_filename=None, manifest_format=None, dry=False):
+                 poster_filename=None, manifest_format=None, dry=False, debug=False):
         self.dry = dry
+        self.debug = debug
         self.poster_size = poster_size or self.DEFAULT_POSTER_SIZE
         self.poster_filename = poster_filename or self.DEFAULT_POSTER_FILENAME
         self.manifest_format = manifest_format or self.DEFAULT_MANIFEST_FORMAT
@@ -73,12 +80,6 @@ class TmdbScrapper:
         # Entry point from TMDb API to download medias
         self.secure_base_url = _api_infos.images["secure_base_url"]
 
-    def get_provider(self):
-        """
-        Shortcut to get the TV provider object from API client.
-        """
-        return TV()
-
     def get_poster_url(self, path):
         """
         Build URL to download poster image from API using its entrypoint and size
@@ -89,27 +90,31 @@ class TmdbScrapper:
             path
         ])
 
-    def serialize_tv_payload(self, provider, tv_id):
+    def store_debug_payload(self, tmdb_id, payload):
+        """
+        Store fetched payload from TmdbScrapper as a JSON file in current working
+        directory, the file will be named with its TMDB ID.
+
+        This is mostly for debug purpose and to enable manually in code when needed.
+        """
+        path = Path("fetch_samples/{}.json".format(tmdb_id))
+        path.write_text(
+            json.dumps(payload, indent=4, cls=ExtendedJsonEncoder)
+        )
+        return path
+
+    def serialize_tv_payload(self, tmdb_id):
         """
         Get informations payload for given TV ID.
         """
         # Fetch payload from API
-        payload = provider.details(tv_id)
+        payload = TV().details(tmdb_id)
 
-        # print()
-        # print("- id:", tv_id)
-        # print("- name:", payload.name)
-        # print("- status:", payload.status)
-        # print("- episode_run_time:", payload.episode_run_time)
-        # print("- first_air_date:", payload.first_air_date)
-        # print("- number_of_episodes:", payload.number_of_episodes)
-        # print("- number_of_seasons:", payload.number_of_seasons)
-        # print("- poster_path:", payload.poster_path)
-        # print("- genres:", [item["name"] for item in payload.genres])
-        # print("- poster_url:", self.get_poster_url(payload.poster_path))
+        if self.debug:
+            self.store_debug_payload(tmdb_id, payload)
 
         return {
-            "tmdb_id": tv_id,
+            "tmdb_id": tmdb_id,
             "tmdb_type": "tv",
             "title": payload.name,
             "status": payload.status,
@@ -117,7 +122,43 @@ class TmdbScrapper:
             "first_air_date": payload.first_air_date,
             "number_of_seasons": payload.number_of_seasons,
             "number_of_episodes": payload.number_of_episodes,
+            "original_language": payload.original_language,
+            "overview": payload.overview,
             "genres": [item["name"] for item in payload.genres],
+            "casting": [
+                [item["name"], item["character"]] for item in payload.credits.cast
+            ],
+            "crew": [
+                [item["name"], item["job"]] for item in payload.credits.crew
+            ],
+        }
+
+    def serialize_movie_payload(self, tmdb_id):
+        """
+        Get informations payload for given MOVIE ID.
+        """
+        # Fetch payload from API
+        payload = Movie().details(tmdb_id)
+
+        if self.debug:
+            self.store_debug_payload(tmdb_id, payload)
+
+        return {
+            "tmdb_id": tmdb_id,
+            "tmdb_type": "movie",
+            "title": payload.title,
+            "status": payload.status,
+            "poster_path": payload.poster_path,
+            "release_date": payload.release_date,
+            "original_language": payload.original_language,
+            "overview": payload.overview,
+            "genres": [item["name"] for item in payload.genres],
+            "casting": [
+                [item["name"], item["character"]] for item in payload.casts.cast
+            ],
+            "crew": [
+                [item["name"], item["job"]] for item in payload.casts.crew
+            ],
         }
 
     def fetch_poster(self, path, basepath):
@@ -172,16 +213,16 @@ class TmdbScrapper:
 
         return diff_lines
 
-    def fetch_tv(self, directory, tv_id, write_diff=False):
+    def fetch_tv(self, directory, tmdb_id, write_diff=False):
         """
         Get informations payload and medias for given TV ID.
 
         This downloads media files and build a YAML manifest to the given directory.
-        """
-        provider = self.get_provider()
 
-        # Fetch and serialize TV show informations
-        data = self.serialize_tv_payload(provider, tv_id)
+        DEPRECATED: In profit of 'fetch_media()' for the new scrapper.
+        """
+        # Fetch and serialize media informations
+        data = self.serialize_tv_payload(tmdb_id)
 
         # Download possible poster image file in destination directory
         fetched_poster = None
@@ -194,6 +235,50 @@ class TmdbScrapper:
             manifest = directory / "manifest.json"
         else:
             manifest = directory / "manifest.yaml"
+
+        diff = self.write_manifest(manifest, data, write_diff=write_diff)
+
+        return (
+            data,
+            manifest,
+            fetched_poster,
+            diff,
+        )
+
+    def fetch_media(self, directory, tmdb_id, tmdb_type="tv", write_diff=False):
+        """
+        Get informations payload and images for given TMDB ID.
+
+        This downloads images files and build a YAML manifest to the given directory.
+        """
+        # TODO: Here we should open the original manifest (if any) to find the option
+        # which would define if the manifest is locked or not. If locked we should not
+        # proceed to fetch payload and let the original manifest unchanged.
+        # As a sample, the manifest is already opened and parsed from method
+        # 'write_manifest()'
+
+        # Fetch and serialize media informations
+        if tmdb_type == "tv":
+            data = self.serialize_tv_payload(tmdb_id)
+        elif tmdb_type == "movie":
+            data = self.serialize_movie_payload(tmdb_id)
+        else:
+            raise NotImplementedError("Given tmdb_type is not implemented: {}".format(
+                tmdb_type
+            ))
+
+        # Download possible poster image file in destination directory
+        fetched_poster = None
+        if data.get("poster_path", None):
+            poster_path = data.pop("poster_path")
+            fetched_poster = self.fetch_poster(poster_path, directory)
+
+        # Build manifest file to destination directory
+        if self.manifest_format == "json":
+            manifest = directory / "manifest.json"
+        else:
+            manifest = directory / "manifest.yaml"
+
         diff = self.write_manifest(manifest, data, write_diff=write_diff)
 
         return (